refactor: share ensure result type
