+++ common/src/types/resources.rs
@@ -133,6 +133,27 @@ pub enum CreateMode {
     Ensure,
 }
 
+/// Result of an ensure operation.
+///
+/// Indicates whether the resource was newly created or an existing resource was
+/// made to match the requested config. Both variants hold the resource's current state.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum EnsureResult<T> {
+    /// Resource was newly created.
+    Created(T),
+    /// Resource already existed and now matches the requested config.
+    Updated(T),
+}
+
+impl<T> EnsureResult<T> {
+    /// Unwrap the inner value regardless of variant.
+    pub fn into_inner(self) -> T {
+        match self {
+            Self::Created(t) | Self::Updated(t) => t,
+        }
+    }
+}
+
 pub static REQUEST_TOKEN_HEADER: http::HeaderName =
     http::HeaderName::from_static("s2-request-token");
