+++ lite/src/handlers/v1/basins.rs
@@ -9,14 +9,11 @@ use s2_common::{
     types::{
         basin::{BasinName, ListBasinsRequest},
         config::{BasinConfig, BasinReconfiguration},
-        resources::{CreateMode, Page, RequestToken},
+        resources::{CreateMode, EnsureResult, Page, RequestToken},
     },
 };
 
-use crate::{
-    backend::{Backend, EnsureResult},
-    handlers::v1::error::ServiceError,
-};
+use crate::{backend::Backend, handlers::v1::error::ServiceError};
 
 pub fn router() -> axum::Router<Backend> {
     use axum::routing::{delete, get, patch, post, put};
